Move Sdk from org.robolectric.internal to ...pluginapi.
