+++ pluginapi/src/main/java/org/robolectric/pluginapi/SdkPicker.java
@@ -3,7 +3,6 @@ package org.robolectric.pluginapi;
 import java.util.List;
 import javax.annotation.Nonnull;
 import org.robolectric.annotation.Config;
-import org.robolectric.internal.Sdk;
 
 public interface SdkPicker {
