+++ pluginapi/src/main/java/org/robolectric/pluginapi/SdkProvider.java
@@ -1,7 +1,6 @@
 package org.robolectric.pluginapi;
 
 import java.util.Collection;
-import org.robolectric.internal.Sdk;
 
 /**
  * A provider of known instances of {@link Sdk}.
