+++ robolectric/src/test/java/org/robolectric/util/TestUtil.java
@@ -11,8 +11,8 @@ import org.junit.runners.model.InitializationError;
 import org.robolectric.LegacyDependencyResolver;
 import org.robolectric.R;
 import org.robolectric.RobolectricTestRunner;
-import org.robolectric.internal.Sdk;
 import org.robolectric.internal.dependency.DependencyResolver;
+import org.robolectric.pluginapi.Sdk;
 import org.robolectric.pluginapi.SdkProvider;
 import org.robolectric.plugins.DefaultSdkProvider;
 import org.robolectric.res.Fs;
